更新短信sdk申请签名参数

See merge request volcengine/volc-sdk-java!312

+++ volc-sdk-java/src/main/java/com/volcengine/model/beans/SmsSignatureInfo.java
@@ -26,7 +26,7 @@ public class SmsSignatureInfo {
     @JSONField(name = "Reason")
     String reason;
     @JSONField(name = "CreatedTime")
-    int createdTime;
+    long createdTime;
     @JSONField(name = "ShortUrlConfig")
     ShortUrlConfig shortUrlConfig;
 }
